requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0

# MCP (Model Context Protocol) dependencies
websockets>=10.0

# Stock data and analysis
yfinance>=0.2.0
alpha-vantage>=2.3.0
ta>=0.10.0  # Technical analysis library
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21

# Data processing and utilities
pydantic>=1.10.0
click>=8.0.0
rich>=12.0.0
tqdm>=4.64.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Caching and storage
redis>=4.0.0
sqlalchemy>=1.4.0

# Configuration and logging
pyyaml>=6.0
structlog>=22.0.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2022.1

# JSON and data serialization
orjson>=3.6.0
ujson>=5.0.0

# HTTP client
httpx>=0.24.0

# Environment and configuration
pydantic-settings>=2.0.0 